Added the data and output folder.
Unpack this into your working directory. One file is index. txt. It has just a single number in it. If you like, you can reset the number to 1. Doesn't matter. The other thing is a folder. As you can see, there are a bunch of Latin and German book titles in it in the first column, beginning in the second row. That is where you put your Latin titles to be lemmatized. Each time you run the code, a data output folder will appear next to the data folder. This part of the code is not done obviously.